requirements.txt
request
pandas
streamlit